backend/requirements.txt
fastapi==0.115.11
uvicorn[standard]==0.34.0
pydantic==2.10.4
python-dotenv==1.0.1
requests==2.32.3
